Factor out register method test (w/o DHT) (#277)
* Factor out register method test (w/o DHT)

* Move skipped test as well

* Residual commit

* Factor out verify method (w/o DHT) (#279)

* Factor out delete tests (w/o DHT)

* Factor out verify method (w/o DHT)

* Factor out delete tests (w/o DHT) (#278)
