+++ tests/index.test.js
@@ -3,13 +3,15 @@ const testList = require('./test-list')
 const testGetMethod = require('./test-get-method')
 const testFollowMethod = require('./test-follow-method')
 const testDeregisterMethod = require('./test-deregister-method')
+const testRegisterMethod = require('./test-register-method')
+const testDeleteMethod = require('./test-delete-method')
+const testVerifyMethod = require('./test-verify-method')
 
 const {
   existsSync,
   promises: { writeFile, readdir, stat, copyFile }
 } = require('fs')
 const { join } = require('path')
-const execa = require('execa')
 const once = require('events.once')
 const test = require('tape')
 const tempy = require('tempy')
@@ -30,7 +32,10 @@ testSdk()
 testList()
 testGetMethod()
 testFollowMethod()
+testRegisterMethod()
 testDeregisterMethod()
+testVerifyMethod()
+testDeleteMethod()
 
 test('sdk re-start', async t => {
   await localDHT()
@@ -158,172 +163,6 @@ test('saveItem: should throw ValidationError with invalid metadata', async t => 
   t.end()
 })
 
-test.skip('register - local contents', async t => {
-  const p2p = createSdk()
-
-  const { rawJSON: profile } = await p2p.init({
-    type: 'profile',
-    title: 'Professor X'
-  })
-  const { rawJSON: content1 } = await p2p.init({
-    type: 'content',
-    title: 'demo',
-    description: 'lorem ipsum'
-  })
-
-  const authors = [encode(profile.url)]
-
-  // update author on content module
-  await p2p.set({ url: content1.url, authors })
-
-  // manually writing a dummy file
-  /*
-  await writeFile(
-    join(p2p.baseDir, encode(content1.url), 'file.txt'),
-    'hola mundo'
-  )
-  */
-  const { metadata: m1 } = await p2p.get(content1.url)
-
-  await p2p.addFiles(content1.url, './tests/testfile.bin')
-
-  await p2p.set({
-    url: content1.url,
-    main: 'testfile.bin'
-  })
-
-  const { metadata } = await p2p.get(content1.url)
-  console.log({ m1version: m1.version })
-  console.log({ m2version: metadata.version })
-  const contentKeyVersion = `${encode(content1.url)}+${metadata.version}`
-
-  try {
-    await p2p.register(contentKeyVersion, encode(profile.url))
-  } catch (err) {
-    t.fail(err.message)
-  }
-  const { rawJSON } = await p2p.get(profile.url)
-  t.same(
-    rawJSON.contents,
-    [contentKeyVersion],
-    'registration results in the addition of a dat key to the contents property of the target profile'
-  )
-  await p2p.destroy()
-  t.end()
-})
-
-test('register - list versions', async t => {
-  const p2p = createSdk({ persist: true })
-
-  const { rawJSON: profile } = await p2p.init({
-    type: 'profile',
-    title: 'Professor X'
-  })
-  const { rawJSON: content1 } = await p2p.init({
-    type: 'content',
-    title: 'demo',
-    description: 'lorem ipsum'
-  })
-
-  const versionsInitial = await p2p.getAllVersions(content1.url)
-  t.same(versionsInitial.length, 0, 'No versions created yet')
-
-  const authors = [encode(profile.url)]
-
-  // update author on content module
-  await p2p.set({ url: content1.url, authors })
-
-  // manually writing a dummy file
-  await writeFile(
-    join(p2p.baseDir, encode(content1.url), 'file.txt'),
-    'hola mundo'
-  )
-
-  await p2p.set({
-    url: content1.url,
-    main: 'file.txt'
-  })
-
-  const { metadata } = await p2p.get(content1.url)
-  const contentKeyVersion = `${encode(content1.url)}+${metadata.version}`
-
-  try {
-    await p2p.register(contentKeyVersion, encode(profile.url))
-  } catch (err) {
-    t.fail(err.message)
-  }
-  const { rawJSON } = await p2p.get(profile.url)
-  t.same(
-    rawJSON.contents,
-    [contentKeyVersion],
-    'registration results in the addition of a dat key to the contents property of the target profile'
-  )
-
-  const versionsFinal = await p2p.getAllVersions(content1.url)
-  t.same(versionsFinal.length, 1, '1 version created')
-
-  await p2p.destroy()
-  t.end()
-})
-
-test('register, restart and list contents', async t => {
-  const dir = tempy.directory()
-  const p2p = new SDK({
-    baseDir: dir,
-    disableSwarm: true
-  })
-  // create profile
-  const { rawJSON: profile } = await p2p.init({
-    type: 'profile',
-    title: 'Professor X'
-  })
-  // create content
-  const { rawJSON: content1 } = await p2p.init({
-    type: 'content',
-    title: 'demo',
-    description: 'lorem ipsum'
-  })
-
-  const authors = [encode(profile.url)]
-  // update author on content module
-  await p2p.set({ url: content1.url, authors })
-  // manually writing a dummy file
-  await writeFile(
-    join(p2p.baseDir, encode(content1.url), 'file.txt'),
-    'hola mundo'
-  )
-  await p2p.set({
-    url: content1.url,
-    main: 'file.txt'
-  })
-  const { metadata } = await p2p.get(content1.url)
-  const contentKeyVersion = `${encode(content1.url)}+${metadata.version}`
-  try {
-    // do the register
-    await p2p.register(contentKeyVersion, encode(profile.url))
-  } catch (err) {
-    t.fail(err.message)
-  }
-  const { rawJSON } = await p2p.get(profile.url)
-  t.same(
-    rawJSON.contents,
-    [contentKeyVersion],
-    'registration results in the addition of a dat key to the contents property of the target profile'
-  )
-  await p2p.destroy()
-
-  const code = join(__dirname, 'childProcessListContent.js')
-  const { stdout, stderr, message, exitCode } = await execa.node(code, [dir])
-
-  if (exitCode !== 0) {
-    console.log(message)
-    t.fail(stderr)
-  }
-  t.same(Number(stdout), 1, 'Expect only one content module')
-
-  t.end()
-})
-
 test('seed and register', async t => {
   const p2p = createSdk({
     swarm: true,
@@ -409,68 +248,6 @@ test('seed and register', async t => {
   t.end()
 })
 
-test('verify', async t => {
-  const p2p = createSdk()
-  const sampleData = [
-    {
-      type: 'content',
-      title: 'demo',
-      description: 'lorem ipsum'
-    },
-    {
-      type: 'content',
-      title: 'demo 2'
-    },
-    { type: 'profile', title: 'Professor X' }
-  ]
-
-  const { rawJSON: profile } = await p2p.init(sampleData[2])
-  const { rawJSON: content1 } = await p2p.init(sampleData[0])
-  const { rawJSON: content2 } = await p2p.init(sampleData[1])
-
-  const authors = [encode(profile.url)]
-
-  // manually writing a dummy file
-  await writeFile(
-    join(p2p.baseDir, encode(content1.url), 'file.txt'),
-    'hola mundo'
-  )
-
-  await p2p.set({
-    url: content1.url,
-    main: 'file.txt'
-  })
-  // update author on content module
-  await p2p.set({ url: content1.url, authors })
-  // update content in author profile
-
-  const { metadata: metadata2 } = await p2p.get(content1.url)
-  const versionedKey = `${encode(content1.url)}+${metadata2.version}`
-
-  try {
-    await p2p.register(versionedKey, profile.url)
-  } catch (err) {
-    t.fail(err.message)
-  }
-
-  const result = await p2p.verify(versionedKey)
-
-  t.ok(result, 'content meets the verification requirements')
-
-  try {
-    await p2p.verify(content2.url)
-  } catch (err) {
-    t.same(
-      err.message,
-      'Module can not be verified: unversioned content',
-      'verify should throw with an unversioned content module'
-    )
-  }
-
-  await p2p.destroy()
-  t.end()
-})
-
 test('verify multiple authors', async t => {
   const p2p = createSdk({
     swarm: true,
@@ -537,52 +314,6 @@ test('verify multiple authors', async t => {
 
   await p2p.destroy()
   await p2p2.destroy()
-  t.end()
-})
-
-test('delete a module from local db', async t => {
-  const dir = tempy.directory()
-  const p2p = new SDK({
-    disableSwarm: true,
-    watch: false,
-    baseDir: dir
-  })
-
-  const modules = await p2p.list()
-  t.equal(modules.length, 0, 'Modules list is empty')
-  // create content
-  const { rawJSON: content } = await p2p.init({
-    type: 'content',
-    title: 'demo',
-    description: 'lorem ipsum'
-  })
-
-  const { rawJSON: content2 } = await p2p.init({
-    type: 'content',
-    title: 'demo 2',
-    description: 'lorem ipsum 2'
-  })
-
-  const modules2 = await p2p.list()
-  t.equal(modules2.length, 2, 'Modules list contains 2 elements')
-  // soft delete content
-  await p2p.delete(content.url)
-
-  const modules3 = await p2p.list()
-  t.equal(modules3.length, 1, 'Modules list contains 1 element')
-
-  // hard delete
-  await p2p.delete(content2.url, true)
-  const baseDir = await readdir(join(p2p.baseDir))
-
-  t.notok(
-    baseDir.includes(encode(content2.url)),
-    'Module folder has been removed'
-  )
-
-  const modules4 = await p2p.list()
-  t.equal(modules4.length, 0, 'Modules list is empty again')
-  await p2p.destroy()
   t.end()
 })
 
@@ -748,78 +479,6 @@ test('delete registered module', async t => {
 
   await p2p.destroy()
   await p2p2.destroy()
-  t.end()
-})
-
-test('delete versioned module', async t => {
-  const dir = tempy.directory()
-  const p2p = new SDK({
-    disableSwarm: true,
-    watch: false,
-    baseDir: dir
-  })
-
-  const modules = await p2p.list()
-  t.equal(modules.length, 0, 'Modules list is empty')
-
-  // create content
-  const { rawJSON: content } = await p2p.init({
-    type: 'content',
-    title: 'demo',
-    description: 'lorem ipsum'
-  })
-
-  const { rawJSON: profile } = await p2p.init({
-    type: 'profile',
-    title: 'professor X',
-    description: 'd'
-  })
-
-  const contentModules = await p2p.listContent()
-  t.equal(contentModules.length, 1, '1 content module exists')
-
-  // register
-  const authors = [encode(profile.url)]
-  // manually writing a dummy file
-  await writeFile(join(dir, encode(content.url), 'file.txt'), 'hola mundo')
-
-  const { metadata } = await p2p.set({
-    url: content.url,
-    authors,
-    main: 'file.txt'
-  })
-
-  const moduleVersioned = `${encode(content.url)}+${metadata.version}`
-  try {
-    await p2p.register(moduleVersioned, encode(profile.url))
-  } catch (err) {
-    t.fail(err.message)
-  }
-  const { rawJSON: updatedProfile } = await p2p.get(profile.url)
-  t.same(updatedProfile.contents.length, 1, 'content registered')
-
-  // hard delete
-  try {
-    await p2p.delete(moduleVersioned, true)
-    t.fail('versioned keys should throw')
-  } catch (err) {
-    t.ok(
-      err.code === 'only_unversioned',
-      'only unversioned keys can be deleted'
-    )
-  }
-
-  const baseDir = await readdir(join(p2p.baseDir))
-
-  const contentModulesFinal = await p2p.listContent()
-  t.equal(contentModulesFinal.length, 1, 'content list remains the same')
-
-  t.ok(
-    baseDir.includes(moduleVersioned),
-    'Module folder has not been removed (deleteFiles)'
-  )
-
-  await p2p.destroy()
   t.end()
 })
